Moved RunProteinFdr + WriteBlibOutput bodies into MergeNodeTask
* MergeNodeTask owns the merge-node phase end to end: 2nd-pass FDR
  sidecar persistence, run-wide protein FDR, and BiblioSpecLite
  output. Nothing on AnalysisPipeline is needed for the merge node
  anymore
* RunProteinFdr / WriteBlibOutput become private instance methods
  on MergeNodeTask; LogInfo / LogWarning calls route through
  _ctx.LogInfo / _ctx.LogWarning where _ctx is the PipelineContext
  set on Run entry. AnalysisPipeline back-reference dropped from
  the constructor
* AnalysisPipeline.cs shrinks by 626 lines (-9% of file)
* Stellar 3-file snapshot regression PASS at every stage

See ai/todos/active/TODO-20260509_osprey_pipeline_tasks.md
